+++ napi/minify/patch.mjs
@@ -8,7 +8,7 @@ data = data.replace(
     `
 if (!nativeBinding && globalThis.process?.versions?.["webcontainer"]) {
   try {
-    nativeBinding = require('./webcontainer-fallback.js');
+    nativeBinding = require('./webcontainer-fallback.cjs');
   } catch (err) {
     loadErrors.push(err)
   }
