+++ napi/minify/webcontainer-fallback.cjs
