+++ napi/transform/webcontainer-fallback.cjs
